Output based TDR
Utilize CEM output features for clustering, in addition to existing input time series of availability, demand, and fuel prices.

run_subperiod_cases.jl: Run individual subperiods independently to obtain individual CEM results (optimal flow) for each subperiod, and stacked together to a full year series to include in clustering. Each subperiod will be weighted to represent the full model hours. This is like solving 1 representative day or week for the year, using the input time series of the particular day or week.

For example, if we are modeling 24 hours in each subperiod for 1 full year, there will be 365 CEMs to run (can be done in parallel, to implement threads in a next update, similar to Benders approach), and each of the 24 hours in each CEM will have a weight of 365 when solved independently. Note that if we use 168 hours in each subperiod, there will be a 53rd subperiod of 24 hours, but it does not have effect in the clustering, since the TDR script will drop any incomplete last subperiods - Only use 8736 hours for clustering in a 8760 hours timseries if modeling representative weeks with 168 hours each.

Users to specify in TDR_settings.json which output feature to include for TDR, at the commodity and asset level, such as VRE and battery electricity edge flow results

In addition, users can choose to turn on policy or not for each subperiod solve.

The output feature flows will be created in a subperiod_results.csv in the system folder with the other time series for clustering, and loaded in the TDR script for parsing and clustering to choose representative weeks.

In addition, a bug in the TDR script is fixed by ensuring that TDR will not run only if required files exists in addition to the folder, or the TDR will run to generate the required files.

Also updated the solar and wind columns such that it does not identify the edges columns in the parsed dataframe when identifying extreme periods for solar and wind.

The new TDR settings include:
   "ClusterSubperiodResults": 1,
    "ClusterSubperiodFileName": "subperiod_results.csv",
    "Zero_Threshold": 0.1, #This is an optional feature to set a threshold that set all output features value to zero if smaller, to prevent very small values in clustering.

    "SubperiodPolicyConstraints": 1,

    "SubperiodResults" : {
        "Commodities" : {
            "Electricity" : {
                "Include": 1,
                "Assets": {
                    "VRE": 1,
                    "Battery": 1,
                    "ThermalPower": 0,
                    "ThermalPowerCCS": 0,
                    "HydroRes": 0,
                    "MustRun": 0
                }
            },
            "Hydrogen" : {
                "Include": 0,
                "Assets": {
                    "Electrolyzer": 0,
                    "ThermalHydrogen": 0,
                    "ThermalHydrogenCCS": 0,
                    "GasStorage": 0
                }
            }
        }
    }